Fix RangeCache#size() bug (#36)
This fixes bug in recent a fix to this method in
011b228f1c70442ba8663bcd8d8b24a3fc5740e6, where the deletion queue was
initially polled instead of peeked, meaning the first entry was
incorrectly deleted unconditionally.

The fix changes to use a ConcurrentSkipListMap rather than
ConcurrentSkipListSet, since the latter does not provide a way to peek
the first element without throwing an exception in the empty case. It
can be done for the map by using its firstEntry() method.